pyproject.toml
[project]
name = "202507-lipari"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.9.0",
    "jupyter>=1.1.1",
    "numpy>=2.2.6",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "tqdm>=4.67.1",
    "weaviate-client>=4.16.1",
]
